// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-ai-usage-summary-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminAiUsageBreakdownHttpResponse } from './admin-ai-usage-breakdown-http-response';
import { AdminAiUsageDailyHttpResponse } from './admin-ai-usage-daily-http-response';
import { AdminAiUsageUserHttpResponse } from './admin-ai-usage-user-http-response';

export interface AdminAiUsageSummaryHttpResponse {
    byDay?: Array<AdminAiUsageDailyHttpResponse>;
    byModel?: Array<AdminAiUsageBreakdownHttpResponse>;
    byOperation?: Array<AdminAiUsageBreakdownHttpResponse>;
    byUser?: Array<AdminAiUsageUserHttpResponse>;
    inputTokens?: number;
    outputTokens?: number;
    totalTokens?: number;
}