// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-ai-usage-user-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminAiUsageUserHttpResponse {
    email?: string | null;
    id?: string;
    inputTokens?: number;
    outputTokens?: number;
    totalTokens?: number;
}